Update utils_py2.py
Changed "bytes_to_uint" to support zero length bit string.  The ANS.1 specification J2735_201603DA.ASN and field vehicle.  The binary stream contained the field vehicle with a zero length.
Example of decoded json output:
"laneType": [
   "vehicle", 
     [
       null, 
       0
     ]
]
As a result the binary stream would not decode.  It is clear if this an issue with the stream, the spec, or the code.  But this change to the code permitted the binary stream to decode completely.  
As a side note: the same ASN spec and binary stream "decoded fine" using http://asn1-playground.oss.com.